+++ R/scatter_plotly.R
@@ -31,6 +31,8 @@
 #' @param add_hover Optional vector of column names in `loc$data` to add to the
 #'   plotly hover text for scatter points.
 #' @param showlegend Logical whether to show a legend for the scatter points.
+#' @param height Height in pixels (optional, defaults to automatic sizing).
+#' @param webGL Logical whether to use webGL or SVG for scatter plot.
 #' @return A `plotly` scatter plot.
 #' @seealso [locus()] [locus_plotly()]
 #' @importFrom plotly add_trace plotly_build
@@ -53,7 +55,9 @@ scatter_plotly <- function(loc,
                            eqtl_gene = NULL,
                            beta = NULL,
                            add_hover = NULL,
-                           showlegend = TRUE) {
+                           showlegend = TRUE,
+                           height = NULL,
+                           webGL = TRUE) {
   if (!inherits(loc, "locus")) stop("Object of class 'locus' required")
   if (is.null(loc$data)) stop("No SNPs/data points", call. = FALSE)
   
@@ -105,6 +109,7 @@ scatter_plotly <- function(loc,
     data$size <- 1L
     data$size[!ind] <- 2L
     sizes <- if (sum(!ind) == 0) c(50, 50) else c(50, 100)
+    if (!webGL) sizes <- sizes/2
     leg <- list(traceorder = "reversed")
   } else {
     if (is.null(eqtl_gene)) {
@@ -145,6 +150,7 @@ scatter_plotly <- function(loc,
                 line = list(width = 1, color = '#999999', dash = 'dash'),
                 x0 = 0, x1 = 1, y0 = -log10(pcutoff), y1 = -log10(pcutoff),
                 xref = "paper", layer = "below")
+  type <- if (webGL) "scattergl" else "scatter"
   
   if (!recomb) {
     if (is.null(beta)) {
@@ -157,8 +163,8 @@ scatter_plotly <- function(loc,
                    text = hovertext, hoverinfo = 'text',
                    key = data[, loc$labs],
                    showlegend = showlegend,
-                   source = "plotly_locus",
-                   type = "scattergl", mode = "markers")
+                   source = "plotly_locus", height = height,
+                   type = type, mode = "markers")
     } else {
       # beta shapes
       p <- plot_ly(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
@@ -170,8 +176,8 @@ scatter_plotly <- function(loc,
                    text = hovertext, hoverinfo = 'text',
                    key = data[, loc$labs],
                    showlegend = showlegend,
-                   source = "plotly_locus",
-                   type = "scattergl", mode = "markers")
+                   source = "plotly_locus", height = height,
+                   type = type, mode = "markers")
     }
     p <- p %>%
       plotly::layout(xaxis = list(title = xlab,
@@ -189,14 +195,14 @@ scatter_plotly <- function(loc,
     ylim2 <- c(-2, 102)
     if (is.null(beta)) {
       # standard plotly
-      p <- plot_ly(source = "plotly_locus") %>%
+      p <- plot_ly(source = "plotly_locus", height = height) %>%
         # recombination line
         add_trace(x = loc$recomb$start / 1e6, y = loc$recomb$value,
                   hoverinfo = "none", colors = scheme,  # colors must go here
                   symbols = symbols,
                   name = "recombination", yaxis = "y2",
                   line = list(color = recomb_col, width = 1.5),
-                  mode = "lines", type = "scattergl", showlegend = FALSE) %>%
+                  mode = "lines", type = type, showlegend = FALSE) %>%
         # scatter plot
         add_trace(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
                   color = data$bg,
@@ -205,17 +211,17 @@ scatter_plotly <- function(loc,
                                 line = list(width = 1, color = marker_outline)),
                   text = hovertext, hoverinfo = 'text', key = data[, loc$labs],
                   showlegend = showlegend,
-                  type = "scattergl", mode = "markers")
+                  type = type, mode = "markers")
     } else {
       # beta shapes
-      p <- plot_ly(source = "plotly_locus") %>%
+      p <- plot_ly(source = "plotly_locus", height = height) %>%
         # recombination line
         add_trace(x = loc$recomb$start / 1e6, y = loc$recomb$value,
                   hoverinfo = "none", colors = scheme,  # colors must go here
                   symbols = symbols, sizes = sizes,
                   name = "recombination", yaxis = "y2",
                   line = list(color = recomb_col, width = 1.5),
-                  mode = "lines", type = "scattergl", showlegend = FALSE) %>%
+                  mode = "lines", type = type, showlegend = FALSE) %>%
         # scatter plot
         add_trace(x = data[, loc$pos] / 1e6, y = data[, loc$yvar],
                   color = data$bg,
@@ -225,7 +231,7 @@ scatter_plotly <- function(loc,
                                 line = list(width = 1, color = marker_outline)),
                   text = hovertext, hoverinfo = 'text', key = data[, loc$labs],
                   showlegend = showlegend,
-                  type = "scattergl", mode = "markers")
+                  type = type, mode = "markers")
     }
     p <- p %>%
       plotly::layout(xaxis = list(title = xlab,